feat(portal): expose resolved timeouts in startup log, move sanitizer to loader
- Export aggregatorTimeoutMs and rpcRequestTimeoutMs as module-level
  consts from url_fetcher and rpc_selector. The latter extracts the env
  read out of the RPCSelector constructor; the field still reads from
  the same module-level value.
- Include both timeouts plus idleTimeoutMaxS in the "Portal config"
  startup log so operators can see the actually-resolved values.
- Move SECRET_FIELDS list and sanitizeConfig() into configuration_loader.ts
  next to the zod schema, so the secret-field knowledge can't drift
  from the schema definition.
- .env.example: rename "Override-only ... (prefer YAML)" to
  "(env vars below override YAML)" — the old phrasing read as if the
  runtime preferred YAML. Tweak the priority-list comment to point
  operators at YAML for those keys due to the pipe syntax.
