+++ portal/server/.env.mainnet.example
@@ -5,7 +5,7 @@
 # Env vars below are supported as overrides — if both YAML and env var
 # are set for a field, the env var wins.
 
-# --- Override-only portal config (prefer YAML) ---
+# --- Portal config (env vars below override YAML) ---
 ENABLE_ALLOWLIST=false
 ENABLE_BLOCKLIST=false
 LANDING_PAGE_OID_B36=46f3881sp4r55fc6pcao9t93bieeejl4vr4k2uv8u4wwyx1a93
@@ -15,7 +15,7 @@ ORIGINAL_PACKAGE_ID=0x26eb7ee8688da02c5f671679524e379f0b837a12f1d1d799f255b7eea2
 B36_DOMAIN_RESOLUTION_SUPPORT=true
 
 # Priority URL lists — format: URL|RETRIES|METRIC (lower metric = higher priority).
-# Left commented because the right values are deployment-specific.
+# Due to the complex syntax of these env vars, it is advised to set them via yaml instead.
 # PREMIUM_RPC_URL_LIST=https://fullnode.mainnet.sui.io|2|100
 # RPC_URL_LIST=https://fullnode.mainnet.sui.io|2|100
 # AGGREGATOR_URL_LIST=https://aggregator.walrus-mainnet.walrus.space|3|100
