+++ portal/server/.env.testnet.example
@@ -5,7 +5,7 @@
 # Env vars below are supported as overrides — if both YAML and env var
 # are set for a field, the env var wins.
 
-# --- Override-only portal config (prefer YAML) ---
+# --- Portal config (env vars below override YAML) ---
 ENABLE_ALLOWLIST=false
 ENABLE_BLOCKLIST=false
 LANDING_PAGE_OID_B36=1p3repujoigwcqrk0w4itsxm7hs7xjl4hwgt3t0szn6evad83q
@@ -16,7 +16,7 @@ B36_DOMAIN_RESOLUTION_SUPPORT=true
 PROMETHEUS_EXPORTER_PORT=9184
 
 # Priority URL lists — format: URL|RETRIES|METRIC (lower metric = higher priority).
-# Left commented because the right values are deployment-specific.
+# Due to the complex syntax of these env vars, it is advised to set them via yaml instead.
 # PREMIUM_RPC_URL_LIST=https://fullnode.testnet.sui.io|2|100
 # RPC_URL_LIST=https://fullnode.testnet.sui.io|2|100
 # AGGREGATOR_URL_LIST=https://aggregator.walrus-testnet.walrus.space|3|100
