+++ portal/worker/.env.mainnet.example
@@ -6,13 +6,13 @@
 # Env vars below are supported as overrides — if both YAML and env var
 # are set for a field, the env var wins.
 
-# --- Override-only worker config (prefer YAML) ---
+# --- Worker config (env vars below override YAML) ---
 PORTAL_DOMAIN_NAME_LENGTH="21"
 SUINS_CLIENT_NETWORK="mainnet"
 ORIGINAL_PACKAGE_ID="0x26eb7ee8688da02c5f671679524e379f0b837a12f1d1d799f255b7eea260ad27"
 
 # Priority URL lists — format: URL|RETRIES|METRIC (lower metric = higher priority).
-# Left commented because the right values are deployment-specific.
+# Due to the complex syntax of these env vars, it is advised to set them via yaml instead.
 # RPC_URL_LIST="https://fullnode.mainnet.sui.io|2|100"
 # AGGREGATOR_URL_LIST="https://aggregator.walrus-mainnet.walrus.space|2|100"
