+++ portal/worker/.env.testnet.example
@@ -6,13 +6,13 @@
 # Env vars below are supported as overrides — if both YAML and env var
 # are set for a field, the env var wins.
 
-# --- Override-only worker config (prefer YAML) ---
+# --- Worker config (env vars below override YAML) ---
 PORTAL_DOMAIN_NAME_LENGTH="21"
 SUINS_CLIENT_NETWORK="testnet"
 ORIGINAL_PACKAGE_ID="0xf99aee9f21493e1590e7e5a9aea6f343a1f381031a04a732724871fc294be799"
 
 # Priority URL lists — format: URL|RETRIES|METRIC (lower metric = higher priority).
-# Left commented because the right values are deployment-specific.
+# Due to the complex syntax of these env vars, it is advised to set them via yaml instead.
 # RPC_URL_LIST="https://fullnode.testnet.sui.io|2|100"
 # AGGREGATOR_URL_LIST="https://aggregator.walrus-testnet.walrus.space|2|100"
